fixup "nic/fec: use linux-imx 6.1.20"
imx_lx_emul: remove nvmem mac-address from DTS

The 6.1.20 kernel tries to acquire the FEC's mac-address by querying
the efuse register. As we do not provide access to them or the data
they contain alter the device-tree to force the driver to fallback
to the mac-address information stored by the bootloader (as was done
in the 5.11 kernel's driver).